requirements.txt
# Django 5 (requires Python 3.10+)
# Note: If using Python 3.9, use Django==4.2.26 instead
Django==5.0.1
# DuckDB backend (optional - will fallback to SQLite if unavailable)
# django-duckdb==0.1.0
# duckdb==0.9.2

# Static file serving for production
whitenoise==6.6.0

# Image handling
Pillow==10.2.0

# Environment variable management
python-decouple==3.8

# Gunicorn for production WSGI server
gunicorn==21.2.0

